+++ gibuAndTakePrj/src/com/gnt/breakdown/controller/BreakdownListController.java
@@ -1,0 +1,17 @@
+package com.gnt.breakdown.controller;
+
+import java.io.IOException;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@WebServlet(urlPatterns = "/member/breakList")
+public class BreakdownListController extends HttpServlet {
+	@Override
+	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		req.getRequestDispatcher("/views/user2/donationBreakdownList.jsp").forward(req, resp);
+	}
+}
