+++ gibuAndTakePrj/src/com/gnt/corpStmt/controller/CorpStmtListController.java
@@ -1,0 +1,18 @@
+package com.gnt.corpStmt.controller;
+
+import java.io.IOException;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@WebServlet(urlPatterns = "/member/stmtRe")
+public class CorpStmtListController extends HttpServlet{
+	@Override
+	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		req.getRequestDispatcher("/views/user2/corpStatementRe.jsp").forward(req, resp);
+	}
+	
+}
