+++ gibuAndTakePrj/src/com/gnt/corpStmt/controller/stmtReController.java
@@ -1,4 +1,4 @@
-package com.kh.corpStmt.controller;
+package com.gnt.corpStmt.controller;
 
 import java.io.IOException;
 
@@ -8,11 +8,11 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet(urlPatterns = "/member/stmtRe")
-public class CorpStmtListController extends HttpServlet{
+@WebServlet (urlPatterns = "/member/stmtList")
+public class stmtReController extends HttpServlet{
+
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		req.getRequestDispatcher("/views/user2/corpStatementRe.jsp").forward(req, resp);
+		req.getRequestDispatcher("/views/user2/corpStatementList.jsp").forward(req, resp);
 	}
-	
 }
