+++ gibuAndTakePrj/src/com/gnt/introduce/controller/IntroduceSiteController.java
@@ -1,0 +1,18 @@
+package com.gnt.introduce.controller;
+
+import java.io.IOException;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@WebServlet(urlPatterns = "/introduce/site")
+public class IntroduceSiteController extends HttpServlet{
+	@Override
+	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		req.getRequestDispatcher("/views/introduce/introduceSite.jsp").forward(req, resp);
+		
+	}
+}
