+++ gibuAndTakePrj/src/com/gnt/introduce/controller/IntroduceWayController.java
@@ -1,0 +1,17 @@
+package com.gnt.introduce.controller;
+
+import java.io.IOException;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@WebServlet(urlPatterns = "/introduce/way")
+public class IntroduceWayController extends HttpServlet {
+	@Override
+	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+		req.getRequestDispatcher("/views/introduce/introduceWay.jsp").forward(req, resp);
+	}
+}
